+++ llms/googleai/new.go
@@ -1,3 +1,5 @@
+// package googleai implements a langchaingo provider for Google AI LLMs.
+// See https://ai.google.dev/ for more details.
 package googleai
 
 import (
